*** empty log message ***
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@152 7b491191-dbf0-0310-aff6-d879d4d69008

+++ mime.c
@@ -171,7 +171,7 @@ int mime_list(char *blkdata, struct list *mimelist)
  *
  * returns -1 on failure, # '\n' on succes
  */
-int mime_readheader(char *blkdata, unsigned long *blkidx, struct list *mimelist, unsigned *headersize)
+int mime_readheader(char *blkdata, unsigned long *blkidx, struct list *mimelist, unsigned long *headersize)
 {
   int valid_mime_lines=0,idx,totallines=0;
   unsigned fieldlen,vallen;
